Move About dock into UI package
